requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
notion-client>=2.2.0
playwright>=1.40.0
google-generativeai>=0.8.0
